rpi/requirements.txt
requests
pyjwt
cryptography
bleak